+++ src/main/java/healthcalc/MetabolicMetricsImpl.java
@@ -2,13 +2,14 @@ package healthcalc;
 
 public class MetabolicMetricsImpl implements MetabolicMetrics{
     
-    public float basalMetabolicRate(float weight, int height, char gender, int age) throws Exception {
-        float res = 0;
+    public double basalMetabolicRate(float weight, int height, char gender, int age) throws Exception {
+        
+        double res = 0;
 
         if (gender == 'm'){
-            res = (float)((10*weight)+(6.25*height)-(5*age)+5);
+            res = ((10*weight)+(6.25*height)-(5*age)+5);
         } else if (gender == 'w'){
-            res = (float)((10*weight)+(6.25*height)-(5*age)-161);
+            res = ((10*weight)+(6.25*height)-(5*age)-161);
         } else if (weight <= 0) {
             throw new Exception("Peso no valido");
         } else if (height <= 0) {
